requirements.txt
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.16.0
cryptography==41.0.7
ecdsa==0.18.0
fastapi==0.105.0
greenlet==3.0.2
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.7
starlette==0.27.0
typing_extensions==4.9.0
